Updated repository data types to match primary key types with the relevant change in the service. This eliminates the need to cast parameters inside the service.

+++ api/src/main/java/com/beanbank/api/repository/TransactionsTypeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface TransactionsTypeRepository extends JpaRepository<TransactionsType, Long> {
+public interface TransactionsTypeRepository extends JpaRepository<TransactionsType, Integer> {
 
 }
